requirements.txt
Flask==2.2.2
Jinja2==3.1.2
mysql-connector==2.2.9
pypng==0.20220715.0
PyQRCode==1.2.1
Pyrogram==2.0.41
PySocks==1.7.1
qrcode==7.3.1
TgCrypto==1.2.3